Move tests to tests/tests.rs
This ensures the tests are only using the public API of the crate.
